Revert "we moved the login and fixed username does not exist and passwords don't match"
