+++ src/components/Dashboard/Home/HomeLogin.jsx
@@ -1,9 +1,16 @@
 import React from "react";
+import Avatar from "@material-ui/core/Avatar";
+import Button from "@material-ui/core/Button";
+import TextField from "@material-ui/core/TextField";
+import Paper from "@material-ui/core/Paper";
 import Box from "@material-ui/core/Box";
 import Grid from "@material-ui/core/Grid";
+import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
+import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
 import { Zoom } from "react-slideshow-image";
 import "react-slideshow-image/dist/styles.css";
+import smile2 from "../../../images/smile2.jfif";
 import s4 from "../../../images/s4.jpg";
 import s5 from "../../../images/s5.jpg";
 import s6 from "../../../images/s6.jpg";
@@ -13,6 +20,16 @@ import s9 from "../../../images/s9.jpg";
 import s10 from "../../../images/s10.jpg"
 import { useEffect } from 'react';
 
+function Copyright() {
+  return (
+    <Typography variant="body2" color="textSecondary" align="center">
+      {"©"}
+      2020 Smile Up Charitable Foundation is a 501(c)3 non-profit organization
+      {new Date().getFullYear()}
+      {"."}
+    </Typography>
+  );
+}
 
 const images = [s4, s5, s6, s7, s8, s9, s10];
 
@@ -42,7 +59,9 @@ const Slideshow = () => {
 };
 const useStyles = makeStyles((theme) => ({
   root: {
-    minHeight: "100vh", 
+    minHeight: "100vh",
+  
+   
   },
   image: {
     maxHeight: "100vh", 
@@ -85,8 +104,6 @@ export default function HomeLogin() {
   useEffect(()=> {
     document.body.classList.add('loginBackground')
   },[])
-
- 
 
   return (
     <div className="body">
